+++ src/cpp/tardigrade_hydraMicromorphicDruckerPragerPlasticity.h
@@ -7,8 +7,8 @@
   ******************************************************************************
   */
 
-#ifndef TARDIGRADE_HYDRA_MICROMORPHIC_LINEAR_ELASTICITY_H
-#define TARDIGRADE_HYDRA_MICROMORPHIC_LINEAR_ELASTICITY_H
+#ifndef TARDIGRADE_HYDRA_MICROMORPHIC_DRUCKER_PRAGER_PLASTICITY_H
+#define TARDIGRADE_HYDRA_MICROMORPHIC_DRUCKER_PRAGER_PLASTICITY_H
 
 #define USE_EIGEN
 #include<tardigrade_vector_tools.h>
@@ -368,6 +368,10 @@ namespace tardigradeHydra{
                                        const parameterType alphaMicro = 0.5,
                                        const parameterType alphaMicroGradient = 0.5 );
 
+        variableType weakMac( const variableType &x, const variableType &a );
+
+        variableType weakMac( const variableType &x, const variableType &a, variableType &dmacdx );
+
         /*!
          * The residual for a micromorphic Drucker Prager plasticity model
          */
@@ -386,7 +390,8 @@ namespace tardigradeHydra{
                 using tardigradeHydra::residualBaseMicromorphic::setdRdT;
 
                 residual( hydraBaseMicromorphic *_hydra, const unsigned int &_numEquations, const unsigned int &plasticConfigurationIndex,
-                          const std::vector< unsigned int > &stateVariableIndices, const floatVector &parameters, const floatType integrationParameter = 0.5 )
+                          const std::vector< unsigned int > &stateVariableIndices, const floatVector &parameters, const floatType integrationParameter = 0.5,
+                          const bool useWeakenedMacaulay = false, const floatType weakenedMacaulayParameter=10, const floatType plasticMultiplierBarrierModulus=1000. )
                         : tardigradeHydra::residualBaseMicromorphic( _hydra, _numEquations ){
                     /*!
                      * The main initialization constructor for the Drucker Prager plasticity residual
@@ -397,6 +402,9 @@ namespace tardigradeHydra{
                      * \param &stateVariableIndices: The indices of the plastic state variables
                      * \param &parameters: The parameter vector
                      * \param &integrationParameter: The integration parameter for the function. 0 is explicit, 1 is implicit.
+                     * \param &useWeakenedMacaulay: A flag for whether to use a weakened Macaulay bracket or not (defaults to false)
+                     * \param &weakenedMacaulayParameter: The value of the parameter for the weakened Macaulay bracket (defaults to 10)
+                     * \param &plasticMultiplierBarrierModulus: The barrier modulus to make sure that the plastic multipliers never go negative (defaults to 1000)
                      */
 
                     _plasticConfigurationIndex = plasticConfigurationIndex;
@@ -406,6 +414,12 @@ namespace tardigradeHydra{
                     _stateVariableIndices = stateVariableIndices;
 
                     _integrationParameter = integrationParameter;
+
+                    _useWeakenedMacaulay = useWeakenedMacaulay;
+
+                    _weakenedMacaulayParameter = weakenedMacaulayParameter;
+
+                    _plasticMultiplierBarrierModulus = plasticMultiplierBarrierModulus;
 
                     TARDIGRADE_ERROR_TOOLS_CATCH( extractMaterialParameters( parameters ) );
 
@@ -420,7 +434,19 @@ namespace tardigradeHydra{
 
                 const floatType* getIntegrationParameter( );
 
+                const bool *useWeakenedMacaulay( ){ return &_useWeakenedMacaulay; }
+
+                const floatType *getWeakenedMacaulayParameter( ){ return &_weakenedMacaulayParameter; }
+
+                const floatType *getPlasticMultiplierBarrierModulus( ){ return &_plasticMultiplierBarrierModulus; }
+
             protected:
+
+                bool _useWeakenedMacaulay; //!< Flag for whether to use the weak Macaulay brackets or not
+
+                floatType _weakenedMacaulayParameter; //!< The weakening parameter for the weak Macaulay brackets
+
+                floatType _plasticMultiplierBarrierModulus; //!< The modulus applied to the plastic multipliers to make sure they are never negative
 
                 unsigned int _numPlasticMultipliers; //!< The number of plastic multipliers. Hard coded to 5 but setting as a variable just in case
 
@@ -959,6 +985,8 @@ namespace tardigradeHydra{
                 virtual void setJacobian( ) override;
 
                 virtual void setdRdD( ) override;
+
+                virtual void setdRdT( ) override;
 
             private: